Create sum of products
To prove your exhaustive testbench will find an error, use a force assignment to override the behavioral
code’s sum sometime after 3⁄4 of the test vectors have been applied. Use the `ifdef and `endif compiler
directives to include or exclude the code that forces an error.
In your report, include log printouts from both the error-free and forced error cases. Use $monitoroff and
$monitoron commands to show only the first dozen or so test vectors (and results) and the last dozen or
so for the error free case. For the forced error case, use them to show the first dozen or so test vectors and
about a dozen test vectors before and including the forced error